Client and tests moved to new locations: .\client\ and .\tests\ respectively
