+++ Lykke.Service.LykkeService/client/Lykke.Service.LykkeService.Client/AutofacExtension.cs
