+++ Lykke.Service.LykkeService/client/Lykke.Service.LykkeService.Client/AutorestClientMapper.cs
