+++ Lykke.Service.LykkeService/client/Lykke.Service.LykkeService.Client/ILykkeServiceClient.cs
