+++ Lykke.Service.LykkeService/client/Lykke.Service.LykkeService.Client/Lykke.Service.LykkeService.csproj
@@ -9,7 +9,7 @@
 
   <ItemGroup>
     <PackageReference Include="Autofac.Extensions.DependencyInjection" Version="4.1.0" />
-    <PackageReference Include="Lykke.Common" Version="1.1.69" />
+    <PackageReference Include="Lykke.Common" Version="1.1.76" />
     <PackageReference Include="Microsoft.Rest.ClientRuntime" Version="2.3.8" />
   </ItemGroup>
