+++ Lykke.Service.LykkeService/client/Lykke.Service.LykkeService.Client/LykkeServiceClient.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using Common.Log;
+
+namespace Lykke.Service.LykkeService
+{
+    public class LykkeServiceClient : ILykkeServiceClient, IDisposable
+    {
+        private readonly ILog _log;
+
+        public LykkeServiceClient(string serviceUrl, ILog log)
+        {
+            _log = log;
+        }
+
+        public void Dispose()
+        {
+            //if (_service == null)
+            //    return;
+            //_service.Dispose();
+            //_service = null;
+        }
+    }
+}
