+++ Lykke.Service.LykkeService/client/Lykke.Service.LykkeService.Client/autorest.ps1
