+++ Lykke.Service.LykkeService/client/Lykke.Service.LykkeService.Client/autorest.sh
@@ -1,0 +1,1 @@
+autorest -Input http://localhost:5000/swagger/v1/swagger.json -CodeGenerator CSharp -OutputDirectory AutorestClient -Namespace Lykke.Service.LykkeService.AutorestClient
