+++ Lykke.Service.LykkeService/src/Lykke.Service.LykkeService.AzureRepositories/Lykke.Service.LykkeService.AzureRepositories.csproj
@@ -8,7 +8,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Lykke.AzureStorage" Version="1.2.22" />
+    <PackageReference Include="Lykke.AzureStorage" Version="1.2.24" />
   </ItemGroup>
 
   <ItemGroup>
