+++ Lykke.Service.LykkeService/src/Lykke.Service.LykkeService.Core/Lykke.Service.LykkeService.Core.csproj
@@ -8,7 +8,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Lykke.Common" Version="1.1.69" />
+    <PackageReference Include="Lykke.Common" Version="1.1.76" />
   </ItemGroup>
 
   <ItemGroup>
