+++ Lykke.Service.LykkeService/src/Lykke.Service.LykkeService.Services/Lykke.Service.LykkeService.Services.csproj
@@ -10,7 +10,7 @@
   <ItemGroup>
     <PackageReference Include="CsvHelper" Version="2.16.3" />
     <PackageReference Include="Flurl.Http" Version="1.1.2" />
-    <PackageReference Include="Lykke.Common" Version="1.1.69" />
+    <PackageReference Include="Lykke.Common" Version="1.1.76" />
   </ItemGroup>
 
   <ItemGroup>
