+++ Lykke.Service.LykkeService/tests/Lykke.Service.LykkeService.Tests/Lykke.Service.LykkeService.Tests.csproj
