+++ Lykke.Service.LykkeService/tests/Lykke.Service.LykkeService.Tests/UnitTest1.cs
